+ SyncDomain::activity_cnt_.

+++ Source/SyncDomain.cpp
@@ -8,14 +8,16 @@ namespace Core {
 
 SyncDomain::SyncDomain () :
 	need_schedule_ (false),
-	state_ (State::IDLE)
+	state_ (State::IDLE),
+	activity_cnt_ (0)
 {
-	Scheduler::create_item ();
 }
 
 SyncDomain::~SyncDomain ()
 {
-	Scheduler::delete_item ();
+	assert (!activity_cnt_);
+	if (activity_cnt_)
+		Scheduler::delete_item ();
 }
 
 void SyncDomain::schedule () NIRVANA_NOEXCEPT
@@ -44,6 +46,31 @@ void SyncDomain::schedule () NIRVANA_NOEXCEPT
 	}
 }
 
+inline
+void SyncDomain::activity_end () NIRVANA_NOEXCEPT
+{
+	if (1 == activity_cnt_.decrement ())
+		Scheduler::delete_item ();
+}
+
+SyncDomain::QueueNode* SyncDomain::create_queue_node (QueueNode* next)
+{
+	assert (0 != activity_cnt_);
+	QueueNode* node = static_cast <QueueNode*> (queue_.allocate_node ());
+	activity_cnt_.increment ();
+	node->domain_ = this;
+	node->next_ = next;
+	return node;
+}
+
+void SyncDomain::release_queue_node (QueueNode* node) NIRVANA_NOEXCEPT
+{
+	assert (node);
+	assert (node->domain_ == this);
+	queue_.deallocate_node (node);
+	activity_end ();
+}
+
 void SyncDomain::execute (Word scheduler_error)
 {
 	assert (State::SCHEDULED == state_);
@@ -55,6 +82,7 @@ void SyncDomain::execute (Word scheduler_error)
 	assert (State::RUNNING == state_);
 	state_ = State::IDLE;
 	schedule ();
+	activity_end ();
 }
 
 void SyncDomain::enter (bool ret)
